refactor: 패키지 구조에 맞게 테스트 클래스들의 위치를 조정합니다.
