+++ src/test/java/lotto/model/service/LottoGenerateServiceTest.java
@@ -1,8 +1,7 @@
-package lotto.model;
+package lotto.model.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.model.service.LottoGenerateService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
